create a policy and authorize in post controller

+++ app/Http/Controllers/PostController.php
@@ -5,9 +5,19 @@ namespace App\Http\Controllers;
 use App\Models\Post;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Routing\Controllers\HasMiddleware;
 
-class PostController extends Controller
+class PostController extends Controller implements HasMiddleware
 {
+    public static function middleware(): array
+    {
+        return [
+            new Middleware('auth', except: ['index','show']),
+        ];
+    }
+
     /**
      * Display a listing of the resource.
      */
@@ -59,6 +69,9 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        //Authorizing the action
+        Gate::authorize('modify', $post);
+
         return view('posts.edit', ['post' => $post]);
     }
 
@@ -67,6 +80,9 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        //Authorizing the action
+        Gate::authorize('modify', $post);
+
         //Validate
         $fields = $request->validate([
             'title' => ['required', 'max:255'],
@@ -85,6 +101,9 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        //Authorizing the action
+        Gate::authorize('modify', $post);
+
         //Delete the post
         $post->delete();
